+++ ECommerceAPI/Core/ECommerceAPI.Domain/Entities/Product.cs
@@ -11,8 +11,8 @@ namespace ECommerceAPI.Domain.Entities
     public sealed class Product : BaseEntity
     {
         public string Name { get; set; } = default!;
-        public int Price { get; set; }
-        public long Stock { get; set; }
+        public long Price { get; set; }
+        public int Stock { get; set; }
         public ICollection<Order> Orders { get; set; } = new Collection<Order>();
     }
 }
